second commit

+++ src/Domain/Invoice.java
@@ -25,7 +25,7 @@ public class Invoice {
         return value;
     }
 
-    public void setValue(int value) {
+    public void setValue(double value) {
         this.value = value;
     }
